+++ src/main/java/com/ripple/BE/learning/dto/ConceptListDTO.java
@@ -1,6 +1,6 @@
 package com.ripple.BE.learning.dto;
 
-import com.ripple.BE.learning.domain.Concept;
+import com.ripple.BE.learning.domain.concept.Concept;
 import java.util.List;
 
 public record ConceptListDTO(List<ConceptDTO> conceptList // 개념 리스트
