+++ src/main/java/com/ripple/BE/learning/dto/QuizListDTO.java
@@ -1,6 +1,6 @@
 package com.ripple.BE.learning.dto;
 
-import com.ripple.BE.learning.domain.Quiz;
+import com.ripple.BE.learning.domain.quiz.Quiz;
 import java.util.List;
 
 public record QuizListDTO(List<QuizDTO> quizList) {
