+++ src/main/java/com/ripple/BE/learning/repository/ConceptRepository.java
@@ -1,7 +1,7 @@
 package com.ripple.BE.learning.repository;
 
-import com.ripple.BE.learning.domain.Concept;
-import com.ripple.BE.learning.domain.LearningSet;
+import com.ripple.BE.learning.domain.concept.Concept;
+import com.ripple.BE.learning.domain.learningset.LearningSet;
 import com.ripple.BE.user.domain.type.Level;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
