+++ src/main/java/com/ripple/BE/learning/repository/QuizRepository.java
@@ -1,7 +1,7 @@
 package com.ripple.BE.learning.repository;
 
-import com.ripple.BE.learning.domain.LearningSet;
-import com.ripple.BE.learning.domain.Quiz;
+import com.ripple.BE.learning.domain.learningset.LearningSet;
+import com.ripple.BE.learning.domain.quiz.Quiz;
 import com.ripple.BE.learning.domain.type.Purpose;
 import com.ripple.BE.user.domain.type.Level;
 import java.util.List;
