+++ src/main/java/com/ripple/BE/learning/service/learningset/LearningSetService.java
@@ -1,7 +1,7 @@
 package com.ripple.BE.learning.service.learningset;
 
-import com.ripple.BE.learning.domain.LearningSet;
-import com.ripple.BE.learning.domain.UserLearningSet;
+import com.ripple.BE.learning.domain.learningset.LearningSet;
+import com.ripple.BE.learning.domain.learningset.UserLearningSet;
 import com.ripple.BE.learning.dto.UserLearningSetListDTO;
 import com.ripple.BE.learning.exception.LearningException;
 import com.ripple.BE.learning.exception.errorcode.LearningErrorCode;
